+++ waterloop-site/src/components/RecruitmentForm/JobPostingPage.tsx
@@ -1,6 +1,6 @@
 import React from 'react'
 import JobPosting from './JobPosting'
-import { samplePostingsData } from '../../pages/Recruitment';
+import { samplePostingsData } from '../../static/consts/postings';
 import { useParams } from 'react-router';
 import styled from 'styled-components';
