+++ waterloop-site/src/static/consts/postings.tsx
@@ -1,0 +1,26 @@
+export const samplePostingsData = [
+    {
+        role: "Goose Designer",
+        deadline: "September 26, 2020",
+        term: "FALL 2020",
+        applicationForm: "https://teamwaterloop.ca",
+        subteam: "SOFTWARE",
+        description: "Laboris incididunt officia culpa consequat dolor adipisicing ullamco ipsum ex. Sit ullamco anim fugiat sint velit. Sit tempor magna adipisicing aliqua anim veniam commodo. Nisi do cupidatat ut laborum commodo eu sunt ut sint. In non irure aliquip in culpa officia ex voluptate commodo sit nostrud cupidatat. Magna pariatur qui deserunt commodo. Ad aliqua proident do duis nisi est labore amet reprehenderit amet consectetur est nulla.",
+        tasks: ["In magna magna eu laborum aliqua labore non in adipisicing commodo anim.", "Quis laborum duis aliqua labore quis amet nisi."],
+        requirements: ["Strong grasp of javascript.", "Quis laborum duis aliqua labore quis amet nisi."],
+        interest: ["Front End Web Development", "Quis laborum duis aliqua labore quis amet nisi."],
+
+    },
+    {
+        role: "Goose Engineer",
+        deadline: "September 26, 2020",
+        term: "FALL 2020",
+        applicationForm: "https://teamwaterloop.ca",
+        subteam: "MECHANICAL",
+        description: "Laboris incididunt officia culpa consequat dolor adipisicing ullamco ipsum ex. Sit ullamco anim fugiat sint velit. Sit tempor magna adipisicing aliqua anim veniam commodo. Nisi do cupidatat ut laborum commodo eu sunt ut sint. In non irure aliquip in culpa officia ex voluptate commodo sit nostrud cupidatat. Magna pariatur qui deserunt commodo. Ad aliqua proident do duis nisi est labore amet reprehenderit amet consectetur est nulla.",
+        tasks: ["In magna magna eu laborum aliqua labore non in adipisicing commodo anim.", "Quis laborum duis aliqua labore quis amet nisi."],
+        requirements: ["Strong grasp of javascript.", "Quis laborum duis aliqua labore quis amet nisi."],
+        interest: ["Front End Web Development", "Quis laborum duis aliqua labore quis amet nisi."],
+
+    }
+];
